# ChibiOS: faster millis micros64 implementation