This greatly speeds up our time functions

@andyp1per this needs testing with DShot and BDShot, on F4 and H7

H743, 32 bit timer, old code:
```
micros()    0.0622 usec/call
micros16()  0.0616 usec/call
millis()    0.5189 usec/call
millis16()  0.5702 usec/call
micros64()  0.2969 usec/call
hrt_micros32()  0.2658 usec/call
hrt_micros64()  0.2769 usec/call
hrt_millis32()  0.4887 usec/call
```
H743, 32 bit timer, new code:
```
micros()    0.0627 usec/call
micros16()  0.0616 usec/call
millis()    0.2662 usec/call
millis16()  0.2787 usec/call
micros64()  0.2002 usec/call
hrt_micros32()  0.0722 usec/call
hrt_micros64()  0.1906 usec/call
hrt_millis32()  0.2456 usec/call
hrt_millis64()  0.2460 usec/call
```

matekh743 (16 bit timer) old code:
```
micros()    0.3432 usec/call
micros16()  0.0618 usec/call
millis()    0.5554 usec/call
millis16()  0.6051 usec/call
micros64()  0.3370 usec/call
hrt_micros32()  0.3107 usec/call
hrt_micros64()  0.3137 usec/call
hrt_millis32()  0.5205 usec/call
```
matekh743 (16 bit timer) new code:
```
micros()    0.1915 usec/call
micros16()  0.0622 usec/call
millis()    0.2572 usec/call
millis16()  0.3009 usec/call
micros64()  0.2059 usec/call
hrt_micros32()  0.1862 usec/call
hrt_micros64()  0.1963 usec/call
hrt_millis32()  0.2489 usec/call
hrt_millis64()  0.2468 usec/call
```

f405, old:
```
micros()    0.1004 usec/call
micros16()  0.1264 usec/call
millis()    1.2906 usec/call
millis16()  1.3632 usec/call
micros64()  0.6696 usec/call
hrt_micros32()  0.5704 usec/call
hrt_micros64()  0.6454 usec/call
hrt_millis32()  1.2894 usec/call
```
f405, new:
```
micros()    0.1008 usec/call
micros16()  0.1254 usec/call
millis()    0.6922 usec/call
millis16()  0.7419 usec/call
micros64()  0.4704 usec/call
hrt_micros32()  0.1247 usec/call
hrt_micros64()  0.4474 usec/call
hrt_millis32()  0.6696 usec/call
hrt_millis64()  0.6666 usec/call
```