@tridge 2023-12-18
> Needs rebasing on new bdshot changes

Done