@andyp1per 2023-12-18
I flew briefly using bdshot on Pixhawk6C FMU outputs and also rebuilt iomcu and tested those outputs as well. The FMU outputs I flew with assertions enabled.